Fix first run error caused by mkdir in getPath
The mkdir in getPath gets executed on import whenever this library is used. This causes an error to be thrown when the electron app that uses it is freshly installed, and the electron userData directory is not yet created.

This change moves the mkdir call to the first place the existence of the extensionsStore is actually required.

Fixes #29